Bartending stuff start 🥃🍻 - Cocktail Shaker, this will mix the drinks (no shit) - Beverage Item, these are the items that you'll be able to mix with. They each have their own fluid and if they're alcoholic. - Mixed Drink, this is the cocktail, which is the finished item that will be the result from two or more beverages after being used in the shaker.
Things will probably get renamed due to how I am unable to choose what to name them. The shaker is really broken at the moment, and is not working as it should. absolute coding slop. so I will be rewriting most of it, probably. uh yeah thats most of it.