thats a big commit... 🧹🧱🔨 - Renamed to Pour Decisions. (thanks gibbly) - Hold beverageItem in offhand with shaker in mainhand, right click and it adds the fluid to the shaker. - beverageItems have a bar that shows how much is left in the beverage (fluid)
thats basically most of it without having to go into too much detail. uh well the solutions for all that fluid stuff is not really the best at the moment. Kinda just tripped over my own code, so that will change. Anyways recipes are now easy to make in CocktailRecipes